Reimplement multiclass_nms in a batched fashion.

| Model              | Original Inf time (s/iter) | New Inf time (s/iter) |
|--------------------|----------------------------|-----------------------|
| Faster R-CNN R-50  | 13.7                       | 16.0                  |
| Mask R-CNN R-50    | 9.9                        | 10.9                  |
| Cascade R-CNN R-50 | 11.7                       | 13.7                  |
| Retinanet R-50     | 12.2                       | 16                    |
| FCOS GN R-50       | 13                         | 19.1                  |
| Mask Scoring R-50  | 9.8                        | 10.8                  |